// databrowser/src/domain/customElements/types.ts
export enum GenericRendererElement {
  ARRAY = 'databrowser-render-array',
  DATE = 'databrowser-render-date',
  IMAGE = 'databrowser-render-image',
  JSON = 'databrowser-render-json',
  STATE = 'databrowser-render-state',
  STRING = 'databrowser-render-string',
  TEXT_HIGHLIGHT = 'databrowser-render-text-highlight',
}